requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
diff-match-patch==20230430
diskcache==5.6.3
distro==1.9.0
fastcore==1.7.5
fastlite==0.0.11
feedparser==6.0.11
filelock==3.16.1
flake8==7.1.1
Flask==3.0.3
Flask-Caching==2.3.0
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grep-ast==0.3.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
litellm==1.47.0
markdown-it-py==3.0.0
markdownify==0.11.6
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
multidict==6.1.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.47.0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
playwright==1.46.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyee==11.1.0
pyflakes==3.2.0
Pygments==2.18.0
pypandoc==1.13
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jobspy==1.1.65
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
scipy==1.13.1
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
SQLAlchemy==2.0.34
sqlite-minutils==3.37.0.post3
starlette==0.38.5
tiktoken==0.7.0
tls-client==1.0.1
tokenizers==0.19.1
tqdm==4.66.5
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
Werkzeug==3.0.4
yarl==1.11.1
zipp==3.20.2
